Ensure uchiwa runs with additional sensu group
`uchiwa` user is created as part of `uchiwa` and `sensu` group, but when using chroot with only `--userspec` it will make it run only as part of the `uchiwa` group. This means no access to the `sensu` accessible `/etc/sensu` directory. As such it will not be able to read the configuration file in `/etc/sensu/uchiwa.js`.